fixed hammer.js

+++ src/app/client/pages/canvas/draw/components/board.vue
@@ -39,7 +39,6 @@
           </div>
 
         </div>
-
       </div>
 
     </aside>
@@ -394,7 +393,7 @@ export default {
         key,
         data,
         type,
-        id: Array.isArray(data) ? data : data.id,
+        id: data.id || '',
         opId: this.genKey(),
         time: new Date().getTime()
       }
@@ -458,7 +457,9 @@ export default {
       !opid && this.socket.emit('sync', 'undo', item, this.board._id, this.board._id)
     },
     deleteSelected() {
-      this.drawer.deleteSelected()
+      if (this.canDelete) {
+        this.drawer.deleteSelected()
+      }
     },
     choose(chooseKey, hiddenAction) {
       if (!this.plugins[chooseKey].useInFollowing && this.notPresenter) {
